Crate hyperion_scheduler

Modules

cleanup

futex

ipc

lock

sleep

task

Statics

READY

RUNNING

Functions

idle

init

init this processors scheduling

process

recv

recv_to

rename

running

schedule

spawn a new process running this closure or a function or a task

send

sleep

sleep_until

spawn

spawn_userspace

spawn a new thread in the currently running process

stop

destroy the current thread and switch to another thread

task

yield_now

switch to another thread

yield_now_wait