Struct hyperion_scheduler::task::TaskInner

pub struct TaskInner {
    pub tid: Tid,
    pub process: Arc<Process>,
    pub state: AtomicCell<TaskState>,
    pub user_stack: Mutex<Stack<UserStack>>,
    pub kernel_stack: Mutex<Stack<KernelStack>>,
    pub job: TakeOnce<Box<dyn FnOnce() + Send + 'static>>,
    pub tls: Once<VirtAddr>,
    /* private fields */
}

Fields
tid: Tid

thread id

thread id’s are per process, each process has at least TID 0

process: Arc<Process>

a shared process ref, multiple tasks can point to the same process

state: AtomicCell<TaskState>

task state, ‘is the task waiting or what?’

user_stack: Mutex<Stack<UserStack>>

lazy initialized user-space stack

kernel_stack: Mutex<Stack<KernelStack>>

lazy initialized kernel-space stack, also used when entering kernel-space from a syscall but not from a interrupt each CPU has its privilege stack in TSS, page faults also have their own stacks per CPU

job: TakeOnce<Box<dyn FnOnce() + Send + 'static>>

thread_entry runs this function once, and stops the process after returning

tls: Once<VirtAddr>

a copy of the master TLS for specifically this task

Implementations

impl TaskInner

pub fn init_tls(&self)